Feat: formatDate 함수 utils 함수로 분리#83
