+++ src/components/Admin/Cabinet/AdminCabinetInformationDisplay.tsx
@@ -1,4 +1,5 @@
 import { CabinetInfo, SelectedMultiCabinetsData } from "@/types/CabinetType";
+import { formatDate } from "@/utils/formatDate";
 import CabinetActionButtons from "@/components/Cabinet/CabinetActionButtons";
 import CabinetSVG from "@/icons/cabinet.svg?react";
 
@@ -12,17 +13,6 @@ interface AdminCabinetInformationDisplayProps extends CabinetInfo {
   selectedStatus: string;
   expiredAt: string | null;
 }
-
-const formatDate = (isoString: string | null): string => {
-  if (!isoString) return "날짜 정보 없음";
-  const date = new Date(isoString);
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  };
-  return date.toLocaleDateString("ko-KR", options);
-};
 
 const AdminCabinetInformationDisplay = ({
   selectedBuilding,
