+++ src/components/Cabinet/SelectedCabinetInformation.tsx
@@ -1,4 +1,5 @@
 import { SelectedCabinetInfo } from "@/types/CabinetType";
+import { formatDate } from "@/utils/formatDate";
 import CabinetActionButtons from "@/components/Cabinet/CabinetActionButtons";
 import CabinetInformationDisplay from "@/components/Cabinet/CabinetInformationDisplay";
 import ConfirmModalView from "@/components/ConfirmModalView";
@@ -9,20 +10,10 @@ import CabinetSVG from "@/icons/cabinet.svg?react";
 
 // 선택된 사물함 정보
 interface SelectedCabinetInformationProps extends SelectedCabinetInfo {
+  setExpiredAt: (expiredAt: string | null) => void;
   isMyCabinet: boolean;
+  setIsMyCabinet: (isMine: boolean) => void;
 }
-
-// 날짜 포맷팅 함수
-const formatDate = (isoString: string | null): string => {
-  if (!isoString) return "날짜 정보 없음";
-  const date = new Date(isoString);
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  };
-  return date.toLocaleDateString("ko-KR", options);
-};
 
 const SelectedCabinetInformation = ({
   selectedCabinet,
