+++ src/utils/formatDate.ts
@@ -1,0 +1,11 @@
+// 날짜 포맷팅 함수
+export const formatDate = (isoString: string | null): string => {
+  if (!isoString) return "날짜 정보 없음";
+  const date = new Date(isoString);
+  const options: Intl.DateTimeFormatOptions = {
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+  };
+  return date.toLocaleDateString("ko-KR", options);
+};
